+++ lib/Component/VCalendar.php
@@ -37,6 +37,7 @@ class VCalendar extends VObject\Document {
      * @var array
      */
     static $componentMap = [
+        'VCALENDAR'     => 'Sabre\\VObject\\Component\\VCalendar',
         'VALARM'        => 'Sabre\\VObject\\Component\\VAlarm',
         'VEVENT'        => 'Sabre\\VObject\\Component\\VEvent',
         'VFREEBUSY'     => 'Sabre\\VObject\\Component\\VFreeBusy',
@@ -509,7 +510,7 @@ class VCalendar extends VObject\Document {
                     'node'    => $this,
                 ];
             }
-            if (count(array_unique($componentTypes)) === 0) {
+            if (count($componentTypes) === 0) {
                 $warnings[] = [
                     'level'   => 3,
                     'message' => 'A calendar object on a CalDAV server must have at least 1 component (VTODO, VEVENT, VJOURNAL).',